backend/pyproject.toml
[tool.poetry]
name = "banking-app"
version = "0.7.0"
description = "Project providing the basic functionalities of a banking system."
authors = [
  "Agata-Lasek",
  "raczu <raczu+contact@protonmail.com>"
]
repository = "https://github.com/Agata-Lasek/banking-app"
readme = "README.md"
classifiers = [
  "Topic :: Software Development",
  "Programming Language :: Python :: 3.11"
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
alembic = "^1.13.1"
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.1.0"
python-jose = "^3.3.0"
tenacity = "^8.2.3"
psycopg2 = "^2.9.9"
email-validator = "^2.1.0.post1"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-extra-types = "^2.4.1"
phonenumbers = "^8.13.27"


requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

